move MultipartUploaderMain to src/test
