+++ README.md
@@ -1,2 +1,16 @@
 # discover-publish
 The Pennsieve Dataset Publishing Workflow
+
+## Multipart Copy/Upload
+
+```
+sbt "discover-publish/test:runMain com.pennsieve.publish.MultipartUploaderMain [arguments]"
+```
+Arguments:
+- `--region`
+- `--maxPartSize`
+- `--sourceBucket`
+- `--sourceKey`
+- `--destinationBucket`
+- `--destinationKey`
+
